utils folder restructured
