+++ navlie/lib/__init__.py
@@ -10,6 +10,7 @@ from .states import (
     SO3State,
     SE23State,
     SL3State,
+    IMMState,
     CompositeState,
     MatrixLieGroupState,
     VectorInput,
